# cmd/api/docs/swagger.yaml
definitions:
...
  req.OTPData:
    properties:
      phoneNumber:
        type: string
    type: object
  req.Otpverifier:
    properties:
      phoneNumber:
        type: string
      pin:
        type: string
    required:
    - phoneNumber
    - pin
    type: object
...
  res.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/admin/order:
    get:
      consumes:
      - application/json
      description: admin can view orders
      operationId: view-order
      parameters:
      - description: 'Start date (format: 2006-1-2)'
        in: query
        name: start
        type: string
      - description: 'End date (format: 2006-1-2)'
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order view
      tags:
      - Order
  /api/admin/order/{orderid}:
    post:
      consumes:
      - application/json
      description: admin can view orders
      operationId: admin-order-details
      parameters:
      - description: orderid
        in: path
        name: orderid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order details
      tags:
      - Order
  /api/admin/order/cancelled:
    get:
      consumes:
      - application/json
      description: admin view order cancelled
      operationId: list-order-by-cancelled
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order cancelled
      tags:
      - Order
  /api/admin/order/placed:
    get:
      consumes:
      - application/json
      description: admin list order placed
      operationId: list-order-by-placed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order placed
      tags:
      - Order
...
  /api/user/order/cancel/{orderId}:
    patch:
      consumes:
      - application/json
      description: user cancel order
      operationId: user-cancel-order
      parameters:
      - description: orderId
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order cart
      tags:
      - Order
  /api/user/order/listall:
    get:
      consumes:
      - application/json
      description: user can view all his orders
      operationId: list-all-order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order status
      tags:
      - Order
  /api/user/order/orderall:
    get:
      consumes:
      - application/json
      description: order all cart items
      operationId: order-all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: order cart
      tags:
      - Cart
  /api/user/otp/send:
    post:
      consumes:
      - application/json
      description: Send OTP to use's mobile
      operationId: send-otp
      parameters:
      - description: User mobile number
        in: body
        name: user_mobile
        required: true
        schema:
          $ref: '#/definitions/req.OTPData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Send OTP to user's mobile
      tags:
      - Otp
  /api/user/otp/verify:
    post:
      consumes:
      - application/json
      description: Validate the  OTP sent to use's mobile
      operationId: validate-otp
      parameters:
      - description: OTP sent to user's mobile number
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/req.Otpverifier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Validate the OTP to user's mobile
      tags:
      - Otp
...
swagger: "2.0"